Added check for missing folders

+++ install_config.py
@@ -11,10 +11,19 @@ os.system("sudo pacman -Syu")
 os.system("sudo pacman -S " + package_string)
 
 #foot config install
-os.system("cp -rf foot.ini $HOME/.config/foot")
+if os.path.exists("$HOME/.config/foot"):
+    os.system("cp -rf foot.ini $HOME/.config/foot")
+else:
+    os.system("mkdir $HOME/.config/foot")
+    os.system("cp -rf fuzzel.ini $HOME/.config/foot/")
+
 
 #fuzzel config install
-os.system("cp -rf fuzzel.ini $HOME/.config/fuzzel")
+if os.path.exists("$HOME/.config/fuzzel/"):
+    os.system("cp -rf fuzzel.ini $HOME/.config/fuzzel")
+else:
+    os.system("mkdir $HOME/.config/fuzzel")
+    os.system("cp -rf fuzzel.ini $HOME/.config/fuzzel/")
 
 #waybar config install 
 os.system("cp -rf waybar $HOME/.config")
